requirements.txt
aiomysql==0.1.1
alembic==1.7.1
asgiref==3.8.1
bcrypt==3.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==3.4.7
dnspython==2.6.1
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.12.1
email_validator==2.2.0
fastapi==0.68.1
greenlet==3.1.1
h11==0.16.0
idna==3.10
importlib_resources==6.4.5
lxml==5.4.0
Mako==1.3.10
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.8.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.32.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.15.0
zipp==3.20.2
